Adicionados os componentes retangulo titulo e container dinheiro
