+++ src/components/Modal_Aviso.js
@@ -1,5 +1,6 @@
 import React, {Component} from 'react';
 import '../css/Modal_Aviso.css';
+import RetanguloTitulo from './Retangulo_Titulo';
 
 class ModalAviso extends Component {
 
@@ -13,16 +14,8 @@ class ModalAviso extends Component {
       <div className="backdrop">
         <div className="modalAviso">
 
-          <div className="retanguloTitulo">
-            <img src={require("../img/logo_menor.png")} alt="Logo da empresa." className="logo"/>
-            <div className="pagamentoParaNome">Aviso</div>
-            <input type="image" src={require('../img/shape-copy.png')} alt="Fechar janela." onClick={this.props.onClose} className="fechar" />
-          </div>
+          <RetanguloTitulo titulo={"Aviso"} onClose={this.props.onClose} />
 
-          <div className="containerVoltar">
-              <img src={require("../img/down_menor.png")} alt="Seta para esquerda, voltar." className="left"/>
-              <div className="voltar" onClick={this.props.onClose}>Voltar</div>
-          </div>
           <div className="containerAviso">
             <span>{this.props.aviso}</span>
           </div>
