+++ src/index.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import ReactDOM from 'react-dom';
-import './css/generic.css';
+import './css/Modal.css';
 import './css/index.css';
 import axios from 'axios';
 import UsuLista from './components/Usu_Lista';
@@ -44,7 +44,7 @@ class App extends Component{
         <div>
 
           <div className="navBar">
-            <picture onClick={this.toggleModal}>
+            <picture>
               <source media="(min-width: 768px)" srcSet={require("./img/logo_maior.png")} />
               <img src={require("./img/logo_menor.png")} alt="Logo da empresa." />
             </picture>
